requirements.txt
# 八字命理系统依赖
pyyaml>=6.0
